% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeconfig}
\alias{makeconfig}
\title{Create an initial configuration file}
\usage{
makeconfig(ofile, df, rows.per.split = 1e+06, is.na = "NA")
}
\arguments{
\item{ofile}{file name path of configuration file to create}

\item{df}{dataframe to use as example for file format}

\item{rows.per.split}{Rows per split, usually a very large number}

\item{is.na}{String representing NA in the data}
}
\description{
Thie function is used to generate a sample configuration from a sample of the data.
The configuration file can then be edited further. The sample of the data is obtained by
reading a sample of the dataset using a function like read.table or read.csv. Since
S datasets are aimed are very large datasets, you usually only need to read a few records,
enough for the import type to auto detect column types.
}
\examples{
\dontrun{
 sample <- read.table('tests/extdata/airline.csv',sep=',', header=TRUE, nrows=100)
 makeconfig('example.yaml',sample)
}
}

