% Generated by roxygen2 (4.0.2): do not edit by hand
\name{op_ge}
\alias{op_ge}
\title{Greater Than or Equal}
\usage{
op_ge(v, value)
}
\arguments{
\item{v}{vector}

\item{value}{vlaue to compare}
}
\value{
index
}
\description{
Greater Than or Equal
}
\examples{
index <- op_ge(ds$splitn("2001-01-01",1,'Distance'),1000)
}

