% Generated by roxygen2 (4.0.2): do not edit by hand
\name{op_gt}
\alias{op_gt}
\title{Greater Than}
\usage{
op_gt(v, value)
}
\arguments{
\item{v}{vector}

\item{value}{vlaue to compare}
}
\value{
index
}
\description{
Greater Than
}
\examples{
index <- op_gt(ds$splitn("2001-01-01",1,'Distance'),1000)
}

