% Generated by roxygen2 (4.0.2): do not edit by hand
\name{op_le}
\alias{op_le}
\title{Less Than or Equal}
\usage{
op_le(v, value)
}
\arguments{
\item{v}{vector}

\item{value}{vlaue to compare}
}
\value{
index
}
\description{
Less Than or Equal
}
\examples{
index <- op_le(ds$splitn("2001-01-01",1,'Distance'),1000)
}

