% Generated by roxygen2 (4.0.2): do not edit by hand
\name{op_lt}
\alias{op_lt}
\title{Less Than}
\usage{
op_lt(v, value)
}
\arguments{
\item{v}{vector}

\item{value}{vlaue to compare}
}
\value{
index
}
\description{
Less Than
}
\examples{
index <- op_lt(ds$splitn("2001-01-01",1,'Distance'),1000)
}

