% Generated by roxygen2 (4.0.1): do not edit by hand
\name{open_sdataset}
\alias{open_sdataset}
\title{Open an S data set file}
\usage{
open_sdataset(path)
}
\arguments{
\item{path}{file name path to the s dataset}
}
\value{
SDataframe
}
\description{
Open an S data set file
}
\examples{
\dontrun{
 sdf <- open_sdataset('rscythica/tests/extdata/airline.sds')
}
}

