% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdataframe_key}
\alias{sdataframe_key}
\title{Create unique key for S dataframe and split}
\usage{
sdataframe_key(partition, split)
}
\arguments{
\item{partition}{key}

\item{split}{index}
}
\value{
Key
}
\description{
Create unique key for S dataframe and split
}
\examples{
k- sdataframe_key'2008-01-03',1)
}

