% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sdf_iterator}
\alias{sdf_iterator}
\title{Create an iterator over S Dataframes}
\usage{
sdf_iterator(sdf)
}
\arguments{
\item{sdf}{S Dataframe}
}
\description{
Creates an iteratior that will iterate over all partition and splits
of an S Dataframe
}
\examples{
#it <- sdf_iterator(df)
#c <- nextElem(it)
#s <- foreach(i=sdf_iterator(df),.combine=sum) \%do\% df$split(i$pkey,i$split,19)
}

