% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[.Rcpp_SDataframe}
\alias{[.Rcpp_SDataframe}
\title{Extract part of an S Data Frame}
\usage{
\method{[}{Rcpp_SDataframe}(x, i, j)
}
\arguments{
\item{x}{S Dataframe}

\item{i}{Row index of the form partion:split}

\item{j}{Column}
}
\value{
Vector in that position
}
\description{
Extract part of an S Data Frame
}
\examples{
\dontrun{
v <-  x["2008-01-03:1","AirTime"]
}
}

