% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sview_columns}
\alias{sview_columns}
\title{Add column filter}
\usage{
sview_columns(v, columns)
}
\arguments{
\item{v}{Scythica View}

\item{columns}{List of column names}
}
\value{
view
}
\description{
Add column filter
}
\examples{
view <- sview(views, c("DepTime","ArrTime"))
}

