% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sview_execute_filter}
\alias{sview_execute_filter}
\title{Create bitmap filters}
\usage{
sview_execute_filter(v)
}
\arguments{
\item{v}{Scythica View}
}
\value{
v
}
\description{
Create bitmap filters
}
\examples{
frow <- sview_executea_filter(views)
}

