% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sview_filter}
\alias{sview_filter}
\title{Filter based on condition}
\usage{
sview_filter(v, variable, operation, value)
}
\arguments{
\item{v}{Scythica View}
}
\value{
number of rows
}
\description{
Filter based on condition
}
\examples{
v <- sview_filter(v, "ArrTime", "eq", 24)
}

