% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sview_partitions}
\alias{sview_partitions}
\title{Add partition filter}
\usage{
sview_partitions(v, partitions)
}
\arguments{
\item{v}{Scythica View}

\item{partitions}{List of partition names}
}
\value{
view
}
\description{
Add partition filter
}
\examples{
view <- sview(views, "2008-01-03")
}

