% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sview_partitions_range}
\alias{sview_partitions_range}
\title{Add partition filter by range}
\usage{
sview_partitions_range(v, from, to)
}
\arguments{
\item{v}{Scythica View}

\item{from}{Starting partition}

\item{to}{End partition}
}
\value{
view
}
\description{
Add partition filter by range
}
\examples{
v <- sview_partitions_range(v,c("2008-01-01"),c('2008-01-10'))
}

