% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sview_partitions_regex}
\alias{sview_partitions_regex}
\title{Add partition filter by regex}
\usage{
sview_partitions_regex(v, from, to)
}
\arguments{
\item{v}{Scythica View}

\item{exp}{Regular Expression}
}
\value{
view
}
\description{
Add partition filter by regex
}
\examples{
v <- sview_partitions_regex(v,c("2008-01-0?"))
}

