% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sview_rawrows}
\alias{sview_rawrows}
\title{Return number of raw rows in filter}
\usage{
sview_rawrows(v)
}
\arguments{
\item{v}{Scythica View}
}
\value{
number of raw rows before filtering
}
\description{
Return number of raw rows in filter
}
\examples{
frow <- sview_rawrows(views)
}

