% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sview_rows}
\alias{sview_rows}
\title{Return number of rows in filter}
\usage{
sview_rows(v)
}
\arguments{
\item{v}{Scythica View}
}
\value{
number of rows
}
\description{
Return number of rows in filter
}
\examples{
frow <- sview_rows(views)
}

